consolidate helper functions into shared/helpers
