+++ components/cards/ChatModelCard.tsx
@@ -10,7 +10,8 @@ import MistralAiLogo from '@/assets/logos/mistral.svg'
 import OpenAiLogo from '@/assets/logos/openai.svg'
 import { LinkButton } from '@/components/ui/LinkButton'
 import { Skeleton } from '@/components/ui/Skeleton'
-import { cn, isValidUrl } from '@/lib/utils'
+import { isValidUrl } from '@/convex/shared/helpers'
+import { cn } from '@/lib/utils'
 
 import type { EChatModel } from '@/convex/types'
 import type { StaticImport } from 'next/dist/shared/lib/get-img-props'
