+++ convex/db/threads.ts
@@ -8,8 +8,7 @@ import { internal } from '../_generated/api'
 import { mutation, query } from '../functions'
 import { kvListV, runConfigV, threadFields } from '../schema'
 import { defaultChatInferenceConfig, defaultImageInferenceConfig } from '../shared/defaults'
-import { extractInferenceConfig } from '../shared/helpers'
-import { extractValidUrlsFromText } from '../shared/utils'
+import { extractInferenceConfig, extractValidUrlsFromText } from '../shared/helpers'
 import { emptyPage, generateSlug } from '../utils'
 import { createJob as createJobNext } from '../workflows/jobs'
 import { getMessageEdges } from './messages'
