+++ convex/shared/helpers.ts
@@ -1,3 +1,4 @@
+import type { Doc } from '../_generated/dataModel'
 import type {
   ChatCompletionConfig,
   EMessage,
@@ -35,4 +36,48 @@ export function extractInferenceConfig(inference: InferenceConfig | undefined): 
     textToImageConfig: inference?.type === 'text-to-image' ? inference : null,
     textToAudioConfig: inference?.type === 'sound-generation' ? inference : null,
   }
+}
+
+export function extractJobsDetails(jobs: Doc<'jobs3'>[]) {
+  const active = jobs.filter((job) => job.status === 'active' || job.status === 'pending')
+  const failed = jobs.filter((job) => job.status === 'failed')
+  const failedJobErrors = failed
+    .map((job) => job.stepResults.at(-1)?.error)
+    .filter((err) => err !== undefined)
+
+  return { active, failed, failedJobErrors }
+}
+
+export function isValidUrl(url: string) {
+  try {
+    new URL(url)
+    return true
+  } catch {
+    return false
+  }
+}
+
+export function extractValidUrlsFromText(text: string): URL[] {
+  const urlRegex = /(https?:\/\/[^\s]+)/g
+  const matches = text.match(urlRegex) || []
+  return matches
+    .map((url) => {
+      try {
+        return new URL(url)
+      } catch {
+        return null
+      }
+    })
+    .filter((url): url is URL => url !== null)
+}
+
+export function hasDelimiter(text: string) {
+  return (
+    text.includes('\n') ||
+    text.includes('.') ||
+    text.includes('?') ||
+    text.includes('!') ||
+    text.includes(',') ||
+    text.length > 100
+  )
 }
