+++ convex/shared/utils.ts
@@ -1,6 +1,5 @@
 import { ConvexError } from 'convex/values'
 
-import type { Doc } from '../_generated/dataModel'
 import type { Value } from 'convex/values'
 
 export function env(name: string) {
@@ -59,39 +58,4 @@ export function stringifyValueForError(value: any) {
     }
     return value
   })
-}
-
-export function getMessageJobsDetails(jobs: Doc<'jobs3'>[]) {
-  const active = jobs.filter((job) => job.status === 'active' || job.status === 'pending')
-  const failed = jobs.filter((job) => job.status === 'failed')
-  const failedJobErrors = failed
-    .map((job) => job.stepResults.at(-1)?.error)
-    .filter((err) => err !== undefined)
-
-  return { active, failed, failedJobErrors }
-}
-
-export function hasDelimiter(text: string) {
-  return (
-    text.includes('\n') ||
-    text.includes('.') ||
-    text.includes('?') ||
-    text.includes('!') ||
-    text.includes(',') ||
-    text.length > 100
-  )
-}
-
-export function extractValidUrlsFromText(text: string): URL[] {
-  const urlRegex = /(https?:\/\/[^\s]+)/g
-  const matches = text.match(urlRegex) || []
-  return matches
-    .map((url) => {
-      try {
-        return new URL(url)
-      } catch {
-        return null
-      }
-    })
-    .filter((url): url is URL => url !== null)
 }
