+++ convex/workflows/pipelines/chat.ts
@@ -6,7 +6,8 @@ import * as vb from 'valibot'
 import { internal } from '../../_generated/api'
 import { internalMutation, internalQuery } from '../../functions'
 import { ResourceKey } from '../../lib/valibot'
-import { env, hasDelimiter } from '../../shared/utils'
+import { hasDelimiter } from '../../shared/helpers'
+import { env } from '../../shared/utils'
 import { jobErrorHandling, WorkflowError } from '../helpers'
 import { createJob } from '../jobs'
